+++ lldb/test/API/commands/session/history/TestSessionHistory.py
@@ -1,5 +1,5 @@
 """
-Test the command history mechanism
+Test the session history command
 """
 
 
@@ -10,13 +10,13 @@ from lldbsuite.test.lldbtest import *
 from lldbsuite.test import lldbutil
 
 
-class CommandHistoryTestCase(TestBase):
+class SessionHistoryTestCase(TestBase):
 
     mydir = TestBase.compute_mydir(__file__)
 
     @no_debug_info_test
     def test_history(self):
-        self.runCmd('command history --clear', inHistory=False)
+        self.runCmd('session history --clear', inHistory=False)
         self.runCmd('breakpoint list', check=False, inHistory=True)  # 0
         self.runCmd('register read', check=False, inHistory=True)  # 1
         self.runCmd('apropos hello', check=False, inHistory=True)  # 2
@@ -32,31 +32,31 @@ class CommandHistoryTestCase(TestBase):
         self.runCmd('frame select 1', check=False, inHistory=True)  # 9
 
         self.expect(
-            "command history -s 3 -c 3",
+            "session history -s 3 -c 3",
             inHistory=True,
             substrs=[
                 '3: memory write',
                 '4: log list',
                 '5: disassemble'])
 
-        self.expect("command history -s 3 -e 3", inHistory=True,
+        self.expect("session history -s 3 -e 3", inHistory=True,
                     substrs=['3: memory write'])
 
         self.expect(
-            "command history -s 6 -e 7",
+            "session history -s 6 -e 7",
             inHistory=True,
             substrs=[
                 '6: expression 1',
                 '7: type summary list -w default'])
 
-        self.expect("command history -c 2", inHistory=True,
+        self.expect("session history -c 2", inHistory=True,
                     substrs=['0: breakpoint list', '1: register read'])
 
-        self.expect("command history -e 3 -c 1", inHistory=True,
+        self.expect("session history -e 3 -c 1", inHistory=True,
                     substrs=['3: memory write'])
 
         self.expect(
-            "command history -e 2",
+            "session history -e 2",
             inHistory=True,
             substrs=[
                 '0: breakpoint list',
@@ -64,43 +64,43 @@ class CommandHistoryTestCase(TestBase):
                 '2: apropos hello'])
 
         self.expect(
-            "command history -s 12",
+            "session history -s 12",
             inHistory=True,
             substrs=[
-                '12: command history -s 6 -e 7',
-                '13: command history -c 2',
-                '14: command history -e 3 -c 1',
-                '15: command history -e 2',
-                '16: command history -s 12'])
+                '12: session history -s 6 -e 7',
+                '13: session history -c 2',
+                '14: session history -e 3 -c 1',
+                '15: session history -e 2',
+                '16: session history -s 12'])
 
         self.expect(
-            "command history -s end -c 3",
+            "session history -s end -c 3",
             inHistory=True,
             substrs=[
-                '15: command history -e 2',
-                '16: command history -s 12',
-                '17: command history -s end -c 3'])
+                '15: session history -e 2',
+                '16: session history -s 12',
+                '17: session history -s end -c 3'])
 
         self.expect(
-            "command history -s end -e 15",
+            "session history -s end -e 15",
             inHistory=True,
             substrs=[
-                '15: command history -e 2',
-                '16: command history -s 12',
-                '17: command history -s end -c 3',
-                'command history -s end -e 15'])
+                '15: session history -e 2',
+                '16: session history -s 12',
+                '17: session history -s end -c 3',
+                'session history -s end -e 15'])
 
-        self.expect("command history -s 5 -c 1", inHistory=True,
+        self.expect("session history -s 5 -c 1", inHistory=True,
                     substrs=['5: disassemble'])
 
-        self.expect("command history -c 1 -s 5", inHistory=True,
+        self.expect("session history -c 1 -s 5", inHistory=True,
                     substrs=['5: disassemble'])
 
-        self.expect("command history -c 1 -e 3", inHistory=True,
+        self.expect("session history -c 1 -e 3", inHistory=True,
                     substrs=['3: memory write'])
 
         self.expect(
-            "command history -c 1 -e 3 -s 5",
+            "session history -c 1 -e 3 -s 5",
             error=True,
             inHistory=True,
             substrs=['error: --count, --start-index and --end-index cannot be all specified in the same invocation'])
